requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.5
